fixes sign errors in <string.h>

+++ libc/string/strncat.c
@@ -2,8 +2,8 @@
 
 char *strncat(char * restrict dest, const char * restrict src, size_t bytes)
 {
-	int i = 0;
-	int dest_idx = strlen(dest);
+	size_t i = 0;
+	size_t dest_idx = strlen(dest);
 	while(src[i] != '\0' && i < bytes)
 		dest[dest_idx++] = src[i++];
 	dest[dest_idx] = '\0';
